discord-bot/requirements.txt
discord.py==2.4.0
requests==2.32.3
openai==1.55.3
python-dotenv

# dev
black==24.8.0
flake8==7.1.1
isort===5.13.2
